* fix numerous warnings, including     - unused declaration     - signed / unsigned comparison     - missing headers     - prototype mismatch
* I know, such commits should be done in a branch before ...
everything has been validated !



git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@13336 32a0c1e6-9308-0410-80dd-f9d2257a62c7

+++ src/Libs/bootstrap/bootstrap.c
@@ -604,7 +604,7 @@ substring_type(int n)
          functional_parameters(ft));
   t = make_type(is_type_functional, ft);
 
-  pips_assert("valid arity", gen_length(functional_parameters(ft))==n);
+  pips_assert("valid arity", (int)gen_length(functional_parameters(ft))==n);
 
   return t;
 }
@@ -629,7 +629,7 @@ assign_substring_type(int n)
          functional_parameters(ft));
   t = make_type(is_type_functional, ft);
 
-  pips_assert("valid arity", gen_length(functional_parameters(ft))==n);
+  pips_assert("valid arity", (int)gen_length(functional_parameters(ft))==n);
 
   return t;
 }
